Move the deriving-functions to a separate module
